requirements-dev.txt
-r requirements.txt
nose
mock
coverage
django-debug-toolbar
https://github.com/WoLpH/Cache-Debug-Toolbar/archive/master.zip#egg=cache_toolbar
pip-tools

#path to deployment package
